+++ flamedisx/block_source.py
@@ -1,6 +1,9 @@
+from copy import deepcopy
+import operator
 import typing as ty
 
 import numpy as np
+import sklearn.neighbors
 import tensorflow as tf
 import pandas as pd
 
@@ -471,6 +474,7 @@ class BlockModelSource(fd.Source):
 
     def _annotate(self, _skip_bounds_computation=False):
         d = self.data
+        MC_bound_dimensions = tuple()
         # By going in reverse order through the blocks, we can use the bounds
         # on hidden variables closer to the final signals (easy to compute)
         # for estimating the bounds on deeper hidden variables.
@@ -481,15 +485,102 @@ class BlockModelSource(fd.Source):
                 b.annotate(d)
             elif b.MC_annotate:
                 for dim in b.MC_annotate_dimensions:
-                    if dim not in self.MC_bound_dimensions:
-                        self.MC_bound_dimensions += (dim,)
+                    if dim not in MC_bound_dimensions:
+                        MC_bound_dimensions += (dim,)
 
         self.MC_bounds(self.source_copy, ('s1', 's2', 'r', 'z'), 'energy', 'energies',
-                       (('rates_vs_energy', 1000),), ('photoelectrons_detected', 'electrons_detected'))
+                       (('rates_vs_energy', 1000),), ('photoelectrons_detected', 'electrons_detected'),
+                       MC_bound_dimensions)
 
         for b in self.model_blocks[::-1]:
             if b.post_MC_annotate:
                 b.annotate(d)
+
+    def MC_bounds(self, source_copy_original, kd_tree_observables,
+                  initial_dimension, initial_attribute,
+                  flat_attributes: ty.Tuple[ty.Tuple[str, int]] = (),
+                  small_MC_filter_dimenions = (),
+                  MC_bound_dimensions: ty.Tuple[str] = ()):
+        """"""
+        source_copy = deepcopy(source_copy_original)
+        MC_data = source_copy.simulate(int(1e6))
+
+        df_full = pd.concat([MC_data, self.data])
+
+        observables_scaled = [(lambda x: (df_full[x] - np.mean(df_full[x])) / \
+            np.std(df_full[x]))(x) for x in kd_tree_observables]
+        data_full = np.array(list(zip(*observables_scaled)))
+
+        data = data_full[len(MC_data)::]
+        data_MC = data_full[0:len(MC_data)]
+
+        tree = sklearn.neighbors.KDTree(data_MC)
+        dist, ind = tree.query(data[::], k=10)
+
+        initial_dimension_MC = MC_data[initial_dimension]
+
+        take_nearest_event = []
+
+        for i in range(len(self.data)):
+            initial_dimension_min = self.data.at[i, initial_dimension + '_min'] = min(initial_dimension_MC.iloc[ind[i]])
+            initial_dimension_max = self.data.at[i, initial_dimension + '_max'] = max(initial_dimension_MC.iloc[ind[i]])
+
+            fix_truth_df = self.data.iloc[i]
+
+            setattr(source_copy, initial_attribute, tf.cast(tf.linspace(initial_dimension_min, initial_dimension_max, 1000),
+                                                            fd.float_type()))
+            for flat_attribute in flat_attributes:
+                setattr(source_copy, flat_attribute[0], tf.ones(flat_attribute[1], fd.float_type()))
+            source_copy.setup_copy()
+
+            sufficient_stats = False
+            count = 0
+
+            while(not sufficient_stats):
+                if (count > 10):
+                    sufficient_stats = True
+                    take_nearest_event.append(True)
+                    continue
+
+                MC_data_small = source_copy.simulate(1000, fix_truth=fix_truth_df)
+                for filter_dim in small_MC_filter_dimenions:
+                    for suffix, comp in (('_min', operator.ge), ('_max', operator.le)):
+                        assert (filter_dim + suffix in self.data), \
+                            f"Bounds on filter dimensions {filter_dim} not computed before MC_bounds"
+                        MC_data_small = MC_data_small.loc[comp(MC_data_small[filter_dim],
+                            self.data[filter_dim + suffix].iloc[i])]
+
+                count += 1
+                if (len(MC_data_small) > 10):
+                    sufficient_stats = True
+                    take_nearest_event.append(False)
+
+            if(take_nearest_event[i] == True):
+                continue
+
+            for x in MC_bound_dimensions:
+                mean_x = MC_data_small[x].mean()
+                std_x = MC_data_small[x].std()
+                self.data.at[i, x + '_min'] = np.floor(mean_x - self.max_sigma * std_x)
+                self.data.at[i, x + '_max'] = np.ceil(mean_x + self.max_sigma * std_x)
+
+        data_bounds = data[[i for i, x in enumerate(take_nearest_event) if not x]]
+        df_bounds = self.data.iloc[[i for i, x in enumerate(take_nearest_event) if not x]].reset_index(drop=True)
+        data_no_bounds = data[[i for i, x in enumerate(take_nearest_event) if x]]
+
+        take_nearest_event_indicies = [i for i, x in enumerate(take_nearest_event) if x]
+
+        if(len(take_nearest_event_indicies) > 0):
+            tree = sklearn.neighbors.KDTree(data_bounds)
+            dist, ind = tree.query(data_no_bounds[::], k=1)
+
+        for i in range(len(take_nearest_event_indicies)):
+            for x in MC_bound_dimensions:
+                self.data.at[take_nearest_event_indicies[i], x + '_min'] = df_bounds.at[ind[i][0], x + '_min']
+                self.data.at[take_nearest_event_indicies[i], x + '_max'] = df_bounds.at[ind[i][0], x + '_max']
+
+        for x in MC_bound_dimensions:
+            self.data[x + '_min'] = self.data[x + '_min'].apply(lambda x : x if x > 0 else 0)
 
 
     def mu_before_efficiencies(self, **params):
